Changed file service to use proper path instead of hard coded path.

+++ Services/IFileService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.IO;
 using System.Text.Json;
 using System.Threading.Tasks;
@@ -13,10 +14,15 @@ public interface IFileService
 
 public class FileService : IFileService
 {
-    public string FilePath => "~/.local/share/paints.json";
+    private static string FilePath => Path.Combine(DataDirectory, "stock.json");
 
+    private static string DataDirectory =>
+        Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "paints");
+    
     public async Task SavePaintList(PaintList list)
     {
+        EnsureDirectoryExists();
+        
         await using var fileStream = File.Create(FilePath);
         
         await JsonSerializer.SerializeAsync<PaintList>(fileStream, list);
@@ -24,9 +30,20 @@ public class FileService : IFileService
 
     public async Task<PaintList?> LoadPaintList()
     {
+        EnsureDirectoryExists();
+
+        if (!File.Exists(FilePath))
+            return null;
+        
         await using var fileStream = File.Create(FilePath);
         var value = await JsonSerializer.DeserializeAsync<PaintList>(fileStream);
 
         return value;
+    }
+
+    private static void EnsureDirectoryExists()
+    {
+        if (!Directory.Exists(DataDirectory))
+            Directory.CreateDirectory(DataDirectory);
     }
 }
